+++ module/utility/damage.mjs
@@ -22,6 +22,13 @@
 // Transdimensional, Trigger, Uncontrolled, Usable As Attack,
 // Variable Advantage, and Variable Special Effects.
 
+export function convertToDiceParts(value) {
+    const dice = Math.floor(value / 5);
+    const halfDice = value % 5 >= 2.5 ? 1 : 0;
+    const plus1 = value % 5 < 2.5 && value % 5 > 0 ? 1 : 0;
+    return { dice, halfDice, plus1 };
+}
+
 // Determine DC solely from item/attack
 export function convertToDcFromItem(item, options) {
     let actor = item.actor;